Repot Assets pallet
