+++ frame/assets/src/lib.rs
@@ -130,152 +130,29 @@ pub mod mock;
 #[cfg(test)]
 mod tests;
 
-use sp_std::prelude::*;
+mod extra_mutator;
+pub use extra_mutator::*;
+mod impl_stored_map;
+mod impl_fungibles;
+mod functions;
+mod types;
+pub use types::*;
+
+use sp_std::{prelude::*, borrow::Borrow};
 use sp_runtime::{
 	RuntimeDebug, TokenError, traits::{
 		AtLeast32BitUnsigned, Zero, StaticLookup, Saturating, CheckedSub, CheckedAdd, Bounded,
+		StoredMapError,
 	}
 };
 use codec::{Encode, Decode, HasCompact};
 use frame_support::{ensure, dispatch::{DispatchError, DispatchResult}};
-use frame_support::traits::{Currency, ReservableCurrency, BalanceStatus::Reserved};
+use frame_support::traits::{Currency, ReservableCurrency, BalanceStatus::Reserved, StoredMap};
 use frame_support::traits::tokens::{WithdrawConsequence, DepositConsequence, fungibles};
 use frame_system::Config as SystemConfig;
 
 pub use weights::WeightInfo;
 pub use pallet::*;
-
-type DepositBalanceOf<T> = <<T as Config>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance;
-
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug)]
-pub struct AssetDetails<
-	Balance,
-	AccountId,
-	DepositBalance,
-> {
-	/// Can change `owner`, `issuer`, `freezer` and `admin` accounts.
-	owner: AccountId,
-	/// Can mint tokens.
-	issuer: AccountId,
-	/// Can thaw tokens, force transfers and burn tokens from any account.
-	admin: AccountId,
-	/// Can freeze tokens.
-	freezer: AccountId,
-	/// The total supply across all accounts.
-	supply: Balance,
-	/// The balance deposited for this asset. This pays for the data stored here.
-	deposit: DepositBalance,
-	/// The ED for virtual accounts.
-	min_balance: Balance,
-	/// If `true`, then any account with this asset is given a provider reference. Otherwise, it
-	/// requires a consumer reference.
-	is_sufficient: bool,
-	/// The total number of accounts.
-	accounts: u32,
-	/// The total number of accounts for which we have placed a self-sufficient reference.
-	sufficients: u32,
-	/// The total number of approvals.
-	approvals: u32,
-	/// Whether the asset is frozen for non-admin transfers.
-	is_frozen: bool,
-}
-
-impl<Balance, AccountId, DepositBalance> AssetDetails<Balance, AccountId, DepositBalance> {
-	pub fn destroy_witness(&self) -> DestroyWitness {
-		DestroyWitness {
-			accounts: self.accounts,
-			sufficients: self.sufficients,
-			approvals: self.approvals,
-		}
-	}
-}
-
-/// A pair to act as a key for the approval storage map.
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug)]
-pub struct ApprovalKey<AccountId> {
-	/// The owner of the funds that are being approved.
-	owner: AccountId,
-	/// The party to whom transfer of the funds is being delegated.
-	delegate: AccountId,
-}
-
-/// Data concerning an approval.
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, Default)]
-pub struct Approval<Balance, DepositBalance> {
-	/// The amount of funds approved for the balance transfer from the owner to some delegated
-	/// target.
-	amount: Balance,
-	/// The amount reserved on the owner's account to hold this item in storage.
-	deposit: DepositBalance,
-}
-
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, Default)]
-pub struct AssetBalance<Balance, Extra> {
-	/// The balance.
-	balance: Balance,
-	/// Whether the account is frozen.
-	is_frozen: bool,
-	/// `true` if this balance gave the account a self-sufficient reference.
-	sufficient: bool,
-	/// Additional "sidecar" data, in case some other pallet wants to use this storage item.
-	extra: Extra,
-}
-
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, Default)]
-pub struct AssetMetadata<DepositBalance> {
-	/// The balance deposited for this metadata.
-	///
-	/// This pays for the data stored in this struct.
-	deposit: DepositBalance,
-	/// The user friendly name of this asset. Limited in length by `StringLimit`.
-	name: Vec<u8>,
-	/// The ticker symbol for this asset. Limited in length by `StringLimit`.
-	symbol: Vec<u8>,
-	/// The number of decimals this asset uses to represent one unit.
-	decimals: u8,
-	/// Whether the asset metadata may be changed by a non Force origin.
-	is_frozen: bool,
-}
-
-/// Witness data for the destroy transactions.
-#[derive(Copy, Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug)]
-pub struct DestroyWitness {
-	/// The number of accounts holding the asset.
-	#[codec(compact)]
-	accounts: u32,
-	/// The number of accounts holding the asset with a self-sufficient reference.
-	#[codec(compact)]
-	sufficients: u32,
-	/// The number of transfer-approvals of the asset.
-	#[codec(compact)]
-	approvals: u32,
-}
-
-/// Trait for allowing a minimum balance on the account to be specified, beyond the
-/// `minimum_balance` of the asset. This is additive - the `minimum_balance` of the asset must be
-/// met *and then* anything here in addition.
-pub trait FrozenBalance<AssetId, AccountId, Balance> {
-	/// Return the frozen balance. Under normal behaviour, this amount should always be
-	/// withdrawable.
-	///
-	/// In reality, the balance of every account must be at least the sum of this (if `Some`) and
-	/// the asset's minimum_balance, since there may be complications to destroying an asset's
-	/// account completely.
-	///
-	/// If `None` is returned, then nothing special is enforced.
-	///
-	/// If any operation ever breaks this requirement (which will only happen through some sort of
-	/// privileged intervention), then `melted` is called to do any cleanup.
-	fn frozen_balance(asset: AssetId, who: &AccountId) -> Option<Balance>;
-
-	/// Called when an account has been removed.
-	fn died(asset: AssetId, who: &AccountId);
-}
-
-impl<AssetId, AccountId, Balance> FrozenBalance<AssetId, AccountId, Balance> for () {
-	fn frozen_balance(_: AssetId, _: &AccountId) -> Option<Balance> { None }
-	fn died(_: AssetId, _: &AccountId) {}
-}
 
 #[frame_support::pallet]
 pub mod pallet {
@@ -1356,615 +1233,6 @@ pub mod pallet {
 				Ok(())
 			})?;
 			Ok(())
-		}
-	}
-}
-
-#[derive(Copy, Clone, PartialEq, Eq)]
-struct TransferFlags {
-	/// The debited account must stay alive at the end of the operation; an error is returned if
-	/// this cannot be achieved legally.
-	keep_alive: bool,
-	/// Less than the amount specified needs be debited by the operation for it to be considered
-	/// successful. If `false`, then the amount debited will always be at least the amount
-	/// specified.
-	best_effort: bool,
-	/// Any additional funds debited (due to minimum balance requirements) should be burned rather
-	/// than credited to the destination account.
-	burn_dust: bool,
-}
-
-#[derive(Copy, Clone, PartialEq, Eq)]
-struct DebitFlags {
-	/// The debited account must stay alive at the end of the operation; an error is returned if
-	/// this cannot be achieved legally.
-	keep_alive: bool,
-	/// Less than the amount specified needs be debited by the operation for it to be considered
-	/// successful. If `false`, then the amount debited will always be at least the amount
-	/// specified.
-	best_effort: bool,
-}
-
-impl From<TransferFlags> for DebitFlags {
-	fn from(f: TransferFlags) -> Self {
-		Self {
-			keep_alive: f.keep_alive,
-			best_effort: f.best_effort,
-		}
-	}
-}
-
-// The main implementation block for the module.
-impl<T: Config> Pallet<T> {
-	// Public immutables
-
-	/// Get the asset `id` balance of `who`.
-	pub fn balance(id: T::AssetId, who: impl sp_std::borrow::Borrow<T::AccountId>) -> T::Balance {
-		Account::<T>::get(id, who.borrow()).balance
-	}
-
-	/// Get the total supply of an asset `id`.
-	pub fn total_supply(id: T::AssetId) -> T::Balance {
-		Asset::<T>::get(id).map(|x| x.supply).unwrap_or_else(Zero::zero)
-	}
-
-	fn new_account(
-		who: &T::AccountId,
-		d: &mut AssetDetails<T::Balance, T::AccountId, DepositBalanceOf<T>>,
-	) -> Result<bool, DispatchError> {
-		let accounts = d.accounts.checked_add(1).ok_or(Error::<T>::Overflow)?;
-		let is_sufficient = if d.is_sufficient {
-			frame_system::Pallet::<T>::inc_sufficients(who);
-			d.sufficients += 1;
-			true
-		} else {
-			frame_system::Pallet::<T>::inc_consumers(who).map_err(|_| Error::<T>::NoProvider)?;
-			false
-		};
-		d.accounts = accounts;
-		Ok(is_sufficient)
-	}
-
-	fn dead_account(
-		what: T::AssetId,
-		who: &T::AccountId,
-		d: &mut AssetDetails<T::Balance, T::AccountId, DepositBalanceOf<T>>,
-		sufficient: bool,
-	) {
-		if sufficient {
-			d.sufficients = d.sufficients.saturating_sub(1);
-			frame_system::Pallet::<T>::dec_sufficients(who);
-		} else {
-			frame_system::Pallet::<T>::dec_consumers(who);
-		}
-		d.accounts = d.accounts.saturating_sub(1);
-		T::Freezer::died(what, who)
-	}
-
-	fn can_increase(id: T::AssetId, who: &T::AccountId, amount: T::Balance) -> DepositConsequence {
-		let details = match Asset::<T>::get(id) {
-			Some(details) => details,
-			None => return DepositConsequence::UnknownAsset,
-		};
-		if details.supply.checked_add(&amount).is_none() {
-			return DepositConsequence::Overflow
-		}
-		let account = Account::<T>::get(id, who);
-		if account.balance.checked_add(&amount).is_none() {
-			return DepositConsequence::Overflow
-		}
-		if account.balance.is_zero() {
-			if amount < details.min_balance {
-				return DepositConsequence::BelowMinimum
-			}
-			if !details.is_sufficient && frame_system::Pallet::<T>::providers(who) == 0 {
-				return DepositConsequence::CannotCreate
-			}
-			if details.is_sufficient && details.sufficients.checked_add(1).is_none() {
-				return DepositConsequence::Overflow
-			}
-		}
-
-		DepositConsequence::Success
-	}
-
-	// The consequence of a withdraw.
-	fn can_decrease(
-		id: T::AssetId,
-		who: &T::AccountId,
-		amount: T::Balance,
-		keep_alive: bool,
-	) -> WithdrawConsequence<T::Balance> {
-		use WithdrawConsequence::*;
-		let details = match Asset::<T>::get(id) {
-			Some(details) => details,
-			None => return UnknownAsset,
-		};
-		if details.supply.checked_sub(&amount).is_none() {
-			return Underflow
-		}
-		if details.is_frozen {
-			return Frozen
-		}
-		let account = Account::<T>::get(id, who);
-		if account.is_frozen {
-			return Frozen
-		}
-		if let Some(rest) = account.balance.checked_sub(&amount) {
-			if let Some(frozen) = T::Freezer::frozen_balance(id, who) {
-				match frozen.checked_add(&details.min_balance) {
-					Some(required) if rest < required => return Frozen,
-					None => return Overflow,
-					_ => {}
-				}
-			}
-
-			let is_provider = false;
-			let is_required = is_provider && !frame_system::Pallet::<T>::can_dec_provider(who);
-			let must_keep_alive = keep_alive || is_required;
-
-			if rest < details.min_balance {
-				if must_keep_alive {
-					WouldDie
-				} else {
-					ReducedToZero(rest)
-				}
-			} else {
-				Success
-			}
-		} else {
-			NoFunds
-		}
-	}
-
-	// Maximum `amount` that can be passed into `can_withdraw` to result in a `WithdrawConsequence`
-	// of `Success`.
-	fn reducible_balance(
-		id: T::AssetId,
-		who: &T::AccountId,
-		keep_alive: bool,
-	) -> Result<T::Balance, Error<T>> {
-		let details = match Asset::<T>::get(id) {
-			Some(details) => details,
-			None => return Err(Error::<T>::Unknown),
-		};
-		ensure!(!details.is_frozen, Error::<T>::Frozen);
-
-		let account = Account::<T>::get(id, who);
-		ensure!(!account.is_frozen, Error::<T>::Frozen);
-
-		let amount = if let Some(frozen) = T::Freezer::frozen_balance(id, who) {
-			// Frozen balance: account CANNOT be deleted
-			let required = frozen.checked_add(&details.min_balance).ok_or(Error::<T>::Overflow)?;
-			account.balance.saturating_sub(required)
-		} else {
-			let is_provider = false;
-			let is_required = is_provider && !frame_system::Pallet::<T>::can_dec_provider(who);
-			if keep_alive || is_required {
-				// We want to keep the account around.
-				account.balance.saturating_sub(details.min_balance)
-			} else {
-				// Don't care if the account dies
-				account.balance
-			}
-		};
-		Ok(amount.min(details.supply))
-	}
-
-	/// Make preparatory checks for debiting some funds from an account. Flags indicate requirements
-	/// of the debit.
-	///
-	/// - `amount`: The amount desired to be debited. The actual amount returned for debit may be
-	///   less (in the case of `best_effort` being `true`) or greater by up to the minimum balance
-	///   less one.
-	/// - `keep_alive`: Require that `target` must stay alive.
-	/// - `respect_freezer`: Respect any freezes on the account or token (or not).
-	/// - `best_effort`: The debit amount may be less than `amount`.
-	///
-	/// On success, the amount which should be debited (this will always be at least `amount` unless
-	/// `best_effort` is `true`) together with an optional value indicating the argument which must
-	/// be passed into the `melted` function of the `T::Freezer` if `Some`.
-	///
-	/// If no valid debit can be made then return an `Err`.
-	fn prep_debit(
-		id: T::AssetId,
-		target: &T::AccountId,
-		amount: T::Balance,
-		f: DebitFlags,
-	) -> Result<T::Balance, DispatchError> {
-		let actual = Self::reducible_balance(id, target, f.keep_alive)?
-			.min(amount);
-		ensure!(f.best_effort || actual >= amount, Error::<T>::BalanceLow);
-
-		let conseq = Self::can_decrease(id, target, actual, f.keep_alive);
-		let actual = match conseq.into_result() {
-			Ok(dust) => actual.saturating_add(dust), //< guaranteed by reducible_balance
-			Err(e) => {
-				debug_assert!(false, "passed from reducible_balance; qed");
-				return Err(e.into())
-			}
-		};
-
-		Ok(actual)
-	}
-
-	/// Make preparatory checks for crediting some funds from an account. Flags indicate
-	/// requirements of the credit.
-	///
-	/// - `amount`: The amount desired to be credited.
-	/// - `debit`: The amount by which some other account has been debited. If this is greater than
-	///   `amount`, then the `burn_dust` parameter takes effect.
-	/// - `burn_dust`: Indicates that in the case of debit being greater than amount, the additional
-	///   (dust) value should be burned, rather than credited.
-	///
-	/// On success, the amount which should be credited (this will always be at least `amount`)
-	/// together with an optional value indicating the value which should be burned. The latter
-	/// will always be `None` as long as `burn_dust` is `false` or `debit` is no greater than
-	/// `amount`.
-	///
-	/// If no valid credit can be made then return an `Err`.
-	fn prep_credit(
-		id: T::AssetId,
-		dest: &T::AccountId,
-		amount: T::Balance,
-		debit: T::Balance,
-		burn_dust: bool,
-	) -> Result<(T::Balance, Option<T::Balance>), DispatchError> {
-		let (credit, maybe_burn) = match (burn_dust, debit.checked_sub(&amount)) {
-			(true, Some(dust)) => (amount, Some(dust)),
-			_ => (debit, None),
-		};
-		Self::can_increase(id, &dest, credit).into_result()?;
-		Ok((credit, maybe_burn))
-	}
-
-	/// Increases the asset `id` balance of `beneficiary` by `amount`.
-	///
-	/// This alters the registered supply of the asset and emits an event.
-	///
-	/// Will return an error or will increase the amount by exactly `amount`.
-	fn do_mint(
-		id: T::AssetId,
-		beneficiary: &T::AccountId,
-		amount: T::Balance,
-		maybe_check_issuer: Option<T::AccountId>,
-	) -> DispatchResult {
-		Self::increase_balance(id, beneficiary, amount, |details| -> DispatchResult {
-			if let Some(check_issuer) = maybe_check_issuer {
-				ensure!(&check_issuer == &details.issuer, Error::<T>::NoPermission);
-			}
-			debug_assert!(T::Balance::max_value() - details.supply >= amount, "checked in prep; qed");
-			details.supply = details.supply.saturating_add(amount);
-			Ok(())
-		})?;
-		Self::deposit_event(Event::Issued(id, beneficiary.clone(), amount));
-		Ok(())
-	}
-
-	/// Increases the asset `id` balance of `beneficiary` by `amount`.
-	///
-	/// LOW-LEVEL: Does not alter the supply of asset or emit an event. Use `do_mint` if you need
-	/// that. This is not intended to be used alone.
-	///
-	/// Will return an error or will increase the amount by exactly `amount`.
-	fn increase_balance(
-		id: T::AssetId,
-		beneficiary: &T::AccountId,
-		amount: T::Balance,
-		check: impl FnOnce(&mut AssetDetails<T::Balance, T::AccountId, DepositBalanceOf<T>>) -> DispatchResult,
-	) -> DispatchResult {
-		if amount.is_zero() { return Ok(()) }
-
-		Self::can_increase(id, beneficiary, amount).into_result()?;
-		Asset::<T>::try_mutate(id, |maybe_details| -> DispatchResult {
-			let details = maybe_details.as_mut().ok_or(Error::<T>::Unknown)?;
-
-			check(details)?;
-
-			Account::<T>::try_mutate(id, beneficiary, |t| -> DispatchResult {
-				let new_balance = t.balance.saturating_add(amount);
-				ensure!(new_balance >= details.min_balance, TokenError::BelowMinimum);
-				if t.balance.is_zero() {
-					t.sufficient = Self::new_account(beneficiary, details)?;
-				}
-				t.balance = new_balance;
-				Ok(())
-			})?;
-			Ok(())
-		})?;
-		Ok(())
-	}
-
-	/// Reduces asset `id` balance of `target` by `amount`. Flags `f` can be given to alter whether
-	/// it attempts a `best_effort` or makes sure to `keep_alive` the account.
-	///
-	/// This alters the registered supply of the asset and emits an event.
-	///
-	/// Will return an error and do nothing or will decrease the amount and return the amount
-	/// reduced by.
-	fn do_burn(
-		id: T::AssetId,
-		target: &T::AccountId,
-		amount: T::Balance,
-		maybe_check_admin: Option<T::AccountId>,
-		f: DebitFlags,
-	) -> Result<T::Balance, DispatchError> {
-		let actual = Self::decrease_balance(id, target, amount, f, |actual, details| {
-			// Check admin rights.
-			if let Some(check_admin) = maybe_check_admin {
-				ensure!(&check_admin == &details.admin, Error::<T>::NoPermission);
-			}
-
-			debug_assert!(details.supply >= actual, "checked in prep; qed");
-			details.supply = details.supply.saturating_sub(actual);
-
-			Ok(())
-		})?;
-		Self::deposit_event(Event::Burned(id, target.clone(), actual));
-		Ok(actual)
-	}
-
-	/// Reduces asset `id` balance of `target` by `amount`. Flags `f` can be given to alter whether
-	/// it attempts a `best_effort` or makes sure to `keep_alive` the account.
-	///
-	/// LOW-LEVEL: Does not alter the supply of asset or emit an event. Use `do_burn` if you need
-	/// that. This is not intended to be used alone.
-	///
-	/// Will return an error and do nothing or will decrease the amount and return the amount
-	/// reduced by.
-	fn decrease_balance(
-		id: T::AssetId,
-		target: &T::AccountId,
-		amount: T::Balance,
-		f: DebitFlags,
-		check: impl FnOnce(
-			T::Balance,
-			&mut AssetDetails<T::Balance, T::AccountId, DepositBalanceOf<T>>,
-		) -> DispatchResult,
-	) -> Result<T::Balance, DispatchError> {
-		if amount.is_zero() { return Ok(amount) }
-
-		let actual = Self::prep_debit(id, target, amount, f)?;
-
-		Asset::<T>::try_mutate(id, |maybe_details| -> DispatchResult {
-			let details = maybe_details.as_mut().ok_or(Error::<T>::Unknown)?;
-
-			check(actual, details)?;
-
-			Account::<T>::try_mutate_exists(id, target, |maybe_account| -> DispatchResult {
-				let mut account = maybe_account.take().unwrap_or_default();
-				debug_assert!(account.balance >= actual, "checked in prep; qed");
-
-				// Make the debit.
-				account.balance = account.balance.saturating_sub(actual);
-				*maybe_account = if account.balance < details.min_balance {
-					debug_assert!(account.balance.is_zero(), "checked in prep; qed");
-					Self::dead_account(id, target, details, account.sufficient);
-					None
-				} else {
-					Some(account)
-				};
-				Ok(())
-			})?;
-
-			Ok(())
-		})?;
-
-		Ok(actual)
-	}
-
-	/// Reduces the asset `id` balance of `source` by some `amount` and increases the balance of
-	/// `dest` by (similar) amount.
-	///
-	/// Returns the actual amount placed into `dest`. Exact semantics are determined by the flags
-	/// `f`.
-	///
-	/// Will fail if the amount transferred is so small that it cannot create the destination due
-	/// to minimum balance requirements.
-	fn do_transfer(
-		id: T::AssetId,
-		source: &T::AccountId,
-		dest: &T::AccountId,
-		amount: T::Balance,
-		maybe_need_admin: Option<T::AccountId>,
-		f: TransferFlags,
-	) -> Result<T::Balance, DispatchError> {
-		// Early exist if no-op.
-		if amount.is_zero() {
-			Self::deposit_event(Event::Transferred(id, source.clone(), dest.clone(), amount));
-			return Ok(amount)
-		}
-
-		// Figure out the debit and credit, together with side-effects.
-		let debit = Self::prep_debit(id, &source, amount, f.into())?;
-		let (credit, maybe_burn) = Self::prep_credit(id, &dest, amount, debit, f.burn_dust)?;
-
-		let mut source_account = Account::<T>::get(id, &source);
-
-		Asset::<T>::try_mutate(id, |maybe_details| -> DispatchResult {
-			let details = maybe_details.as_mut().ok_or(Error::<T>::Unknown)?;
-
-			// Check admin rights.
-			if let Some(need_admin) = maybe_need_admin {
-				ensure!(&need_admin == &details.admin, Error::<T>::NoPermission);
-			}
-
-			// Skip if source == dest
-			if source == dest {
-				return Ok(())
-			}
-
-			// Burn any dust if needed.
-			if let Some(burn) = maybe_burn {
-				// Debit dust from supply; this will not saturate since it's already checked in prep.
-				debug_assert!(details.supply >= burn, "checked in prep; qed");
-				details.supply = details.supply.saturating_sub(burn);
-			}
-
-			// Debit balance from source; this will not saturate since it's already checked in prep.
-			debug_assert!(source_account.balance >= debit, "checked in prep; qed");
-			source_account.balance = source_account.balance.saturating_sub(debit);
-
-			Account::<T>::try_mutate(id, &dest, |a| -> DispatchResult {
-				// Calculate new balance; this will not saturate since it's already checked in prep.
-				debug_assert!(a.balance.checked_add(&credit).is_some(), "checked in prep; qed");
-				let new_balance = a.balance.saturating_add(credit);
-
-				// Create a new account if there wasn't one already.
-				if a.balance.is_zero() {
-					a.sufficient = Self::new_account(&dest, details)?;
-				}
-
-				a.balance = new_balance;
-				Ok(())
-			})?;
-
-			// Remove source account if it's now dead.
-			if source_account.balance < details.min_balance {
-				debug_assert!(source_account.balance.is_zero(), "checked in prep; qed");
-				Self::dead_account(id, &source, details, source_account.sufficient);
-				Account::<T>::remove(id, &source);
-			} else {
-				Account::<T>::insert(id, &source, &source_account)
-			}
-
-			Ok(())
-		})?;
-
-		Self::deposit_event(Event::Transferred(id, source.clone(), dest.clone(), credit));
-		Ok(credit)
-	}
-}
-
-impl<T: Config> fungibles::Inspect<<T as SystemConfig>::AccountId> for Pallet<T> {
-	type AssetId = T::AssetId;
-	type Balance = T::Balance;
-
-	fn total_issuance(asset: Self::AssetId) -> Self::Balance {
-		Asset::<T>::get(asset).map(|x| x.supply).unwrap_or_else(Zero::zero)
-	}
-
-	fn minimum_balance(asset: Self::AssetId) -> Self::Balance {
-		Asset::<T>::get(asset).map(|x| x.min_balance).unwrap_or_else(Zero::zero)
-	}
-
-	fn balance(
-		asset: Self::AssetId,
-		who: &<T as SystemConfig>::AccountId,
-	) -> Self::Balance {
-		Pallet::<T>::balance(asset, who)
-	}
-
-	fn reducible_balance(
-		asset: Self::AssetId,
-		who: &<T as SystemConfig>::AccountId,
-		keep_alive: bool,
-	) -> Self::Balance {
-		Pallet::<T>::reducible_balance(asset, who, keep_alive).unwrap_or(Zero::zero())
-	}
-
-	fn can_deposit(
-		asset: Self::AssetId,
-		who: &<T as SystemConfig>::AccountId,
-		amount: Self::Balance,
-	) -> DepositConsequence {
-		Pallet::<T>::can_increase(asset, who, amount)
-	}
-
-	fn can_withdraw(
-		asset: Self::AssetId,
-		who: &<T as SystemConfig>::AccountId,
-		amount: Self::Balance,
-	) -> WithdrawConsequence<Self::Balance> {
-		Pallet::<T>::can_decrease(asset, who, amount, false)
-	}
-}
-
-impl<T: Config> fungibles::Mutate<<T as SystemConfig>::AccountId> for Pallet<T> {
-	fn mint_into(
-		asset: Self::AssetId,
-		who: &<T as SystemConfig>::AccountId,
-		amount: Self::Balance,
-	) -> DispatchResult {
-		Self::do_mint(asset, who, amount, None)
-	}
-
-	fn burn_from(
-		asset: Self::AssetId,
-		who: &<T as SystemConfig>::AccountId,
-		amount: Self::Balance,
-	) -> Result<Self::Balance, DispatchError> {
-		let f = DebitFlags {
-			keep_alive: false,
-			best_effort: false,
-		};
-		Self::do_burn(asset, who, amount, None, f)
-	}
-
-	fn slash(
-		asset: Self::AssetId,
-		who: &<T as SystemConfig>::AccountId,
-		amount: Self::Balance,
-	) -> Result<Self::Balance, DispatchError> {
-		let f = DebitFlags {
-			keep_alive: false,
-			best_effort: true,
-		};
-		Self::do_burn(asset, who, amount, None, f)
-	}
-}
-
-impl<T: Config> fungibles::Transfer<T::AccountId> for Pallet<T> {
-	fn transfer(
-		asset: Self::AssetId,
-		source: &T::AccountId,
-		dest: &T::AccountId,
-		amount: T::Balance,
-		keep_alive: bool,
-	) -> Result<T::Balance, DispatchError> {
-		let f = TransferFlags {
-			keep_alive,
-			best_effort: false,
-			burn_dust: false
-		};
-		Self::do_transfer(asset, source, dest, amount, None, f)
-	}
-}
-
-impl<T: Config> fungibles::Unbalanced<T::AccountId> for Pallet<T> {
-	fn set_balance(_: Self::AssetId, _: &T::AccountId, _: Self::Balance) -> DispatchResult {
-		unreachable!();
-	}
-	fn set_total_issuance(id: T::AssetId, amount: Self::Balance) {
-		Asset::<T>::mutate_exists(id, |maybe_asset| if let Some(ref mut asset) = maybe_asset {
-			asset.supply = amount
-		});
-	}
-	fn decrease_balance(asset: T::AssetId, who: &T::AccountId, amount: Self::Balance)
-		-> Result<Self::Balance, DispatchError>
-	{
-		let f = DebitFlags { keep_alive: false, best_effort: false };
-		Self::decrease_balance(asset, who, amount, f, |_, _| Ok(()))
-	}
-	fn decrease_balance_at_most(asset: T::AssetId, who: &T::AccountId, amount: Self::Balance)
-		-> Self::Balance
-	{
-		let f = DebitFlags { keep_alive: false, best_effort: true };
-		Self::decrease_balance(asset, who, amount, f, |_, _| Ok(()))
-			.unwrap_or(Zero::zero())
-	}
-	fn increase_balance(asset: T::AssetId, who: &T::AccountId, amount: Self::Balance)
-		-> Result<Self::Balance, DispatchError>
-	{
-		Self::increase_balance(asset, who, amount, |_| Ok(()))?;
-		Ok(amount)
-	}
-	fn increase_balance_at_most(asset: T::AssetId, who: &T::AccountId, amount: Self::Balance)
-		-> Self::Balance
-	{
-		match Self::increase_balance(asset, who, amount, |_| Ok(())) {
-			Ok(()) => amount,
-			Err(_) => Zero::zero(),
 		}
 	}
 }
